skills/paddleocr-doc-parsing/scripts/requirements-optimize.txt
# File Optimization Dependencies
# Install with: pip install -r scripts/paddleocr-doc-parsing/requirements-optimize.txt

# Image processing
Pillow>=10.0.0

# PDF processing
pypdfium2>=4.0.0
